Add files via upload
This is the power-law covariance matrix elements added version of JAVELIN. Also I made some trivial improvements such as making the code work when the input data are magnitude time series and able to save histograms when mass trials. And the code of the spear_covfunc.so, spear_covfunc.f90 is lost, but all parts of it can be found in later versions of the spear_covfunc.f90 files.